+++ resources/js/components/FormField.vue
@@ -5,7 +5,7 @@
         <div class="o1-flex o1-flex-wrap o1-items-stretch o1-w-full o1-relative">
           <div class="o1-flex o1--mr-px">
             <span
-              class="o1-flex o1-items-center o1-leading-normal o1-rounded o1-border o1-border-r-0 o1-border-gray-300 o1-px-3 o1-whitespace-no-wrap o1-text-80 o1-text-sm o1-font-bold dark:o1-border-gray-700 o1-rounded-r-none"
+              class="o1-flex o1-items-center o1-leading-normal o1-rounded o1-border o1-border-r-0 o1-border-gray-300 o1-px-3 o1-whitespace-nowrap o1-text-gray-600 o1-text-sm o1-font-bold dark:o1-border-gray-700 dark:o1-text-gray-400 o1-rounded-r-none"
             >
               {{ field.currency }}
             </span>